Move tests for state store committer
